okhttp-issue1903-demo
When requesting tens of files in separate threads, cancelling some results in
an java.io.InterruptedIOException: thread interrupted exception that causes
hasJournalErrors = true;.

Consequently, within private synchronized Editor

    edit(String key, long expectedSequenceNumber)

the cache is no longer updated because of hasJournalErrors.

Once in this state I do not believe that rebuildJournal(); is called until the
cache is reinitialized. I looked in journalRebuildRequired() but that does not
consider the hasJournalErrors flag.

Perhaps I should guard against threads being cancelled that are writing to the
cache? RFC.